// lib/api/generated/models/aiSettingsResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 1.1.2.1+2
 */
import type { AiSettings } from './aiSettings';
import type { AiUsageStatus } from './aiUsageStatus';
import type { GroqFreePlanLimits } from './groqFreePlanLimits';

export interface AiSettingsResponse {
  env_model_fallback: string;
  free_plan_limits?: null | GroqFreePlanLimits;
  groq_key_configured: boolean;
  settings: AiSettings;
  usage: AiUsageStatus;
}